requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=1.0.0
google-adk>=1.0.0
fastmcp>=3.0.0
duckdb>=1.2.0
openpyxl>=3.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
